Remove return statements, separate support.rb classes in files, change example.rb
